requirements.txt
google-genai
python-dotenv
phonenumbers
pyttsx3
SpeechRecognition
pipwin
pyaudio
selenium